# code/week-11/app/src/main/java/com/asprak/todolistd/domain/Category.kt
package com.asprak.todolistd.domain

import kotlinx.serialization.SerialName
import kotlinx.serialization.Serializable

@Serializable
data class Category(
    val id: Int,
    @SerialName("user_id")
    val userId: Int,
    val name: String,
)