# code/week-11/app/src/main/java/com/asprak/todolistd/feature/auth/data/dto/RegisterReq.kt
package com.asprak.todolistd.feature.auth.data.dto

import kotlinx.serialization.Serializable

@Serializable
data class RegisterReq(
    val name: String,
    val email: String,
    val password: String
)